requirements.txt
ultralytics 8.3.163
flask  3.1.1
python-dotenv =0.21.0 
nest_asyncio 1.6.0
python-telegram-bot -22.2


anyio-4.9.0 blinker-1.9.0 click-8.2.1 contourpy-1.3.2 cycler-0.12.1 exceptiongroup-1.3.0 flask-3.1.1 fonttools-4.58.5 h11-0.16.0 httpcore-1.0.9 httpx-0.28.1 itsdangerous-2.2.0 jinja2-3.1.6 kiwisolver-1.4.8 markupsafe-3.0.2 matplotlib-3.10.3 mpmath-1.3.0 nest_asyncio-1.6.0 networkx-3.4.2 nvidia-cublas-cu12-12.6.4.1 nvidia-cuda-cupti-cu12-12.6.80 nvidia-cuda-nvrtc-cu12-12.6.77 nvidia-cuda-runtime-cu12-12.6.77 nvidia-cudnn-cu12-9.5.1.17 nvidia-cufft-cu12-11.3.0.4 nvidia-cufile-cu12-1.11.1.6 nvidia-curand-cu12-10.3.7.77 nvidia-cusolver-cu12-11.7.1.2 nvidia-cusparse-cu12-12.5.4.2 nvidia-cusparselt-cu12-0.6.3 nvidia-nccl-cu12-2.26.2 nvidia-nvjitlink-cu12-12.6.85 nvidia-nvtx-cu12-12.6.77 opencv-python-4.12.0.88 pillow-11.3.0 psutil-7.0.0 py-cpuinfo-9.0.0 pyparsing-3.2.3 python-telegram-bot-22.2 scipy-1.15.3 sniffio-1.3.1 sympy-1.14.0 torch-2.7.1 torchvision-0.22.1 triton-3.3.1 ultralytics-8.3.163 ultralytics-thop-2.0.14 werkzeug-3.1.3